:recycle: [refactor] #32 - refactor structure, 회원가입 시 카트생성
