+++ src/main/java/profect/group1/goormdotcom/common/config/WebMvcConfig.java
@@ -5,7 +5,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import profect.group1.goormdotcom.user.presentation.auth.LoginUserArgumentResolver;
+import profect.group1.goormdotcom.user.controller.auth.LoginUserArgumentResolver;
 
 @Configuration
 @RequiredArgsConstructor
