+++ src/main/java/profect/group1/goormdotcom/user/domain/User.java
@@ -4,7 +4,6 @@ import lombok.Builder;
 import lombok.Getter;
 import java.util.UUID;
 import profect.group1.goormdotcom.user.domain.enums.UserRole;
-import profect.group1.goormdotcom.user.domain.enums.SellerApprovalStatus;
 import java.time.LocalDateTime;
 
 @Getter
@@ -16,5 +15,4 @@ public class User {
     private String name;
     private UserRole role;
     private String brandId;
-    private SellerApprovalStatus sellerApprovalStatus;
 }
