--- src/main/java/profect/group1/goormdotcom/user/domain/enums/SellerApprovalStatus.java
@@ -1,7 +1,0 @@
-package profect.group1.goormdotcom.user.domain.enums;
-
-public enum SellerApprovalStatus {
-    PENDING,
-    APPROVED,
-    REJECTED
-}
