--- src/main/java/profect/group1/goormdotcom/user/infra/converter/UserJpaConverter.java
@@ -1,32 +1,0 @@
-package profect.group1.goormdotcom.user.infra.converter;
-
-import profect.group1.goormdotcom.user.domain.User;
-import profect.group1.goormdotcom.user.infra.UserJpaEntity;
-
-public class UserJpaConverter {
-    public static User toDomain(UserJpaEntity jpa) {
-        return User.builder()
-            .id(jpa.getId())
-            .createdAt(jpa.getCreatedAt())
-            .email(jpa.getEmail())
-            .name(jpa.getName())
-            .role(jpa.getRole())
-            .brandId(jpa.getBrandId())
-            .sellerApprovalStatus(jpa.getSellerApprovalStatus())
-            .build();
-    }
-
-    public static UserJpaEntity toEntity(User domain) {
-        return UserJpaEntity.builder()
-            .id(domain.getId()) 
-            .createdAt(domain.getCreatedAt())
-            .email(domain.getEmail())
-            .name(domain.getName())
-            .role(domain.getRole())
-            .brandId(domain.getBrandId())
-            .sellerApprovalStatus(domain.getSellerApprovalStatus())
-            .build();
-    }
-}
-
-
