--- src/main/java/profect/group1/goormdotcom/user/presentation/auth/LoginUser.java
@@ -1,14 +1,0 @@
-package profect.group1.goormdotcom.user.presentation.auth;
-
-import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.PARAMETER)
-@Retention(RetentionPolicy.RUNTIME)
-@Documented
-public @interface LoginUser {
-
-}
